+++ client/src/main/java/replicant/ConnectorState.java
@@ -26,7 +26,12 @@ public enum ConnectorState
   /**
    * The service is in error state. This error may occur during connection, disconnection or in normal operation.
    */
-  ERROR;
+  ERROR,
+  /**
+   * The service is in a fatal error state. The client should not attempt a retry
+   * until the error has been addressed. This often indicates a security error.
+   */
+  FATAL_ERROR;
 
   /**
    * Return true if state is one of <code>DISCONNECTING</code> or <code>CONNECTING</code>.